% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy.R
\name{build_euclid_dist_matrix}
\alias{build_euclid_dist_matrix}
\title{Build averaged pairwise euclidean distance between blocks}
\usage{
build_euclid_dist_matrix(dt_series)
}
\arguments{
\item{dt_series}{a data.table of series with block_id blocks}
}
\value{
a matrix of averaged pairwise euclidean distance between blocks
}
\description{
Build averaged pairwise euclidean distance between blocks
}
