% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score.R
\name{get_dt_score_avg}
\alias{get_dt_score_avg}
\title{Compute average mean absolute error for all columns}
\usage{
get_dt_score_avg(array_sample, dt_series, col_vec = NULL, ref_block_id = NULL)
}
\arguments{
\item{array_sample}{array of samples to test, one sample by column}

\item{dt_series}{data.table of series}

\item{col_vec}{a character vector with columns names for quantile score computation}
}
\value{
a data.table of scores, one row per sample
}
\description{
Compute average mean absolute error for all columns
}
