% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score.R
\name{get_dt_score_energy}
\alias{get_dt_score_energy}
\title{Compute energy scores}
\usage{
get_dt_score_energy(array_sample, l_mat_A, ref_block_id = NULL)
}
\arguments{
\item{array_sample}{array of samples to test, one sample by column}

\item{l_mat_A}{list of A matrices for the energy scores}
}
\value{
a data.table of scores, one row per sample
}
\description{
Compute energy scores
}
