% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score.R
\name{get_dt_score_quantile}
\alias{get_dt_score_quantile}
\title{Compute and concatenate quantile scores for multiple columns, fast version}
\usage{
get_dt_score_quantile(
  array_sample,
  dt_series,
  col_vec = NULL,
  upper_quantile_boundary = 0.998,
  ref_block_id = NULL
)
}
\arguments{
\item{array_sample}{array of samples to test, one sample by column}

\item{dt_series}{data.table of series}

\item{col_vec}{a character vector with columns names for quantile score computation}

\item{...}{Other arguments passed to internal get_dt_score_quantile_one_col}
}
\value{
a data.table of scores, one row per sample
}
\description{
Compute and concatenate quantile scores for multiple columns, fast version
}
