% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy.R
\name{matrixblockmean}
\alias{matrixblockmean}
\title{Compute block mean of a matrix}
\usage{
matrixblockmean(m, blockszi, blockszj)
}
\arguments{
\item{m}{Matrix}

\item{blockszi}{row block size}

\item{blockszj}{col block size}
}
\value{
an averaged matrix
}
\description{
Compute block mean of a matrix
}
